Move the service names to service_names.py
Also to avoid circular imports
