+++ cfy_manager/components/restservice/db.py
@@ -35,13 +35,13 @@ from ..components_constants import (
     MANAGER_STATUS_REPORTER,
 )
 
-from .. import DATABASE_SERVICE
 from ..service_names import (
     MANAGER,
     POSTGRESQL_CLIENT,
     POSTGRESQL_SERVER,
     RABBITMQ,
     RESTSERVICE,
+    DATABASE_SERVICE
 )
 
 from ... import constants
