+++ cfy_manager/components/service_names.py
@@ -30,3 +30,8 @@ RESTSERVICE = 'restservice'
 SANITY = 'sanity'
 STAGE = 'stage'
 USAGE_COLLECTOR = 'usage_collector'
+
+
+DATABASE_SERVICE = 'database_service'
+QUEUE_SERVICE = 'queue_service'
+MANAGER_SERVICE = 'manager_service'
