+++ cfy_manager/utils/common.py
@@ -35,11 +35,12 @@ from ..logger import get_logger
 from ..exceptions import ProcessExecutionError
 
 from cfy_manager.components.components_constants import SERVICES_TO_INSTALL
-from cfy_manager.components import (QUEUE_SERVICE,
-                                    MANAGER_SERVICE,
-                                    DATABASE_SERVICE)
-from cfy_manager.components.service_names import POSTGRESQL_SERVER
-
+from cfy_manager.components.service_names import (
+    POSTGRESQL_SERVER,
+    QUEUE_SERVICE,
+    MANAGER_SERVICE,
+    DATABASE_SERVICE
+)
 from . import subprocess_preexec
 
 logger = get_logger('utils')
